+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/feat/DependencyPinningBundle.kt
@@ -2,6 +2,8 @@ package fluxo.conf.feat
 
 import fluxo.conf.FluxoKmpConfContext
 import fluxo.conf.data.VersionCatalogConstants.VC_PINNED_BUNDLE_ALIAS
+import fluxo.conf.impl.kotlin.KOTLIN_2_1
+import fluxo.conf.impl.kotlin.KOTLIN_PLUGIN_VERSION_RAW
 import fluxo.conf.impl.logDependency
 import fluxo.log.d
 import fluxo.log.l
@@ -21,6 +23,21 @@ internal fun FluxoKmpConfContext.prepareDependencyPinningBundle() {
     val p = rootProject
 
     val pinnedDeps: PinnedDeps = HashMap()
+
+    if (kotlinPluginVersion >= KOTLIN_2_1) {
+        // KGP doesn't depend on the `kotlin-compiler-embeddable` dependency
+        // starting from Kotlin 2.1.0
+        // Other plugins can bring incompatible versions of the compiler.
+        // https://kotlinlang.slack.com/archives/C0KLZSCHF/p1729256644747559?thread_ts=1729151089.194689&cid=C0KLZSCHF
+        val version = runCatching { KOTLIN_PLUGIN_VERSION_RAW }.getOrNull()
+            ?: kotlinPluginVersion.toString()
+        val compilerEmbeddable = object : ModuleIdentifier {
+            override fun getGroup() = "org.jetbrains.kotlin"
+            override fun getName() = "kotlin-compiler-embeddable"
+        }
+        pinnedDeps[compilerEmbeddable] = Pair(version, "Pinned to Kotlin plugin version")
+    }
+
     val bundleAliases = libs.bundleAliases
     if (bundleAliases.isNotEmpty()) {
         for (alias in bundleAliases) {
@@ -73,7 +90,7 @@ private fun FluxoKmpConfContext.collectPinnedDependencies(
     logger: Logger,
     pinnedDeps: PinnedDeps,
 ) {
-    // Filter "pinned and "pinned.*" bundles
+    // Filter "pinned" and "pinned.*" bundles
     alias.startsWith(ALIAS, ignoreCase = true) && alias.run {
         val l = length
         l == ALIAS.length || l > ALIAS.length && this[ALIAS.length] == '.'